requirements.txt
accelerate
datasets
jsonlines
matplotlib
python-dotenv
scikit-learn
sentence_transformers
torch
transformers
